requirements.txt
fastapi
mysql-connector-python
uvicorn
python-dotenv
passlib
redis
twilio
colorama
jose